Contracts/src/AssetTransfer.test.ts
import { AssetTransfer } from "./AssetTransfer";
import { CreditWorthiness } from "./Creditworthiness";


describe("AssetTransfer Contract", () => {
  let creditWorthinessContract: CreditWorthiness;
  let assetTransferContract: AssetTransfer;

  beforeAll(async () => {
  
    creditWorthinessContract = await deployCreditWorthinessContract();
    // Deploy the AssetTransfer contract with the address of the CreditWorthiness contract
    assetTransferContract = await deployAssetTransferContract(creditWorthinessContract.address);
  });

  it("should transfer an asset if the sender is creditworthy", async () => {
    const senderPublicKey = "creditworthyUserPublicKey";
    const recipientPublicKey = "recipientUserPublicKey";
    const amount = 100; // Asset amount to transfer

    // First, ensure the sender is considered creditworthy
    await creditWorthinessContract.submitCreditScore(senderPublicKey, 700); // High credit score

    const transferResult = await assetTransferContract.transferAssetIfCreditworthy(senderPublicKey, recipientPublicKey, amount);
    expect(transferResult).toBe(true);
  });

  it("should not transfer an asset if the sender is not creditworthy", async () => {
    const senderPublicKey = "nonCreditworthyUserPublicKey";
    const recipientPublicKey = "anotherRecipientUserPublicKey";
    const amount = 100;

    // Sender has a low credit score
    await creditWorthinessContract.submitCreditScore(senderPublicKey, 600); // Low credit score

    const transferResult = await assetTransferContract.transferAssetIfCreditworthy(senderPublicKey, recipientPublicKey, amount);
    expect(transferResult).toBe(false);
  });
});
function beforeAll(arg0: () => Promise<void>) {
  throw new Error("Function not implemented.");
}

function deployCreditWorthinessContract(): any {
  throw new Error("Function not implemented.");
}

function deployAssetTransferContract(address: any): any {
  throw new Error("Function not implemented.");
}

function expect(transferResult: any) {
  throw new Error("Function not implemented.");
}

